Switching the SSH Key Deletion to a defer
